requirements.txt
absl-py==1.4.0
addict==2.4.0
aioice==0.9.0
aiortc==1.5.0
albumentations==1.3.1
altair==5.0.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==22.2.0
Augmentor==0.2.12
av==10.0.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
bing-image-downloader==1.1.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
compreface-sdk==0.6.0
cryptography==41.0.1
cycler==0.10.0
darknet==0.3
dataclasses==0.6
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dlib @ file:///home/conda/feedstock_root/build_artifacts/dlib_1652035094040/work
dnspython==2.3.0
docopt==0.6.2
easydict==1.9
email-validator==2.0.0.post2
entrypoints==0.4
et-xmlfile==1.1.0
ez-setup==0.9
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.99.1
fastjsonschema==2.16.3
filelock==3.12.2
fireflyalgorithm==0.3.4
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.38.0
fsspec==2023.1.0
future==0.18.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.20.0
google-auth-oauthlib==0.4.6
google-crc32c==1.5.0
google-pasta==0.2.0
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
huggingface-hub==0.16.4
idna==2.10
ifaddr==0.2.0
iglovikov-helper-functions==0.0.53
imagecorruptions==1.1.2
imageio==2.31.0
imgaug==0.4.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
imutils==0.5.4
install==1.3.5
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jpeg4py==0.1.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.23.6
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.11.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mysql==0.0.3
mysql-connector-python-rf==2.2.2
mysqlclient==2.1.1
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.5.3
notebook_shim==0.2.2
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1649806299270/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
olefile==0.46
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.1
packaging==23.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.9.0
psutil==5.9.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.10
pydeck==0.8.1b0
pyee==10.0.2
Pygments==2.14.0
pylibsrtp==0.8.0
Pympler==1.0.1
pyodbc==4.0.39
pyOpenSSL==23.2.0
pyparsing==2.4.7
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.6
python-pygaze==0.7.6
python-version==0.0.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==5.4.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
qudida==0.0.4
regex==2023.6.3
requests==2.30.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==13.4.1
rsa==4.9
safetensors==0.3.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.27.0
streamlit==1.23.1
streamlit-webrtc==0.44.7
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.0
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.9.12
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==1.7.0
torchvision==0.4.0+cu92
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
ujson==5.7.0
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.2.3
wget==3.2
widgetsnbextension==4.0.5
wrapt==1.15.0
yarg==0.1.9
zipp==3.15.0
